+++ familyhelper-webapi-node/src/main/java/com/dwarfeng/familyhelper/webapi/node/controller/v1/clannad/AvatarController.java
@@ -1,0 +1,145 @@
+package com.dwarfeng.familyhelper.webapi.node.controller.v1.clannad;
+
+import com.dwarfeng.dutil.basic.io.IOUtil;
+import com.dwarfeng.familyhelper.clannad.sdk.bean.entity.FastJsonAvatarInfo;
+import com.dwarfeng.familyhelper.clannad.stack.bean.dto.AvatarDownloadInfo;
+import com.dwarfeng.familyhelper.clannad.stack.bean.dto.AvatarUploadInfo;
+import com.dwarfeng.familyhelper.clannad.stack.bean.entity.AvatarInfo;
+import com.dwarfeng.familyhelper.webapi.stack.service.clannad.AvatarResponseService;
+import com.dwarfeng.subgrade.sdk.bean.dto.FastJsonResponseData;
+import com.dwarfeng.subgrade.sdk.bean.dto.ResponseDataUtil;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.http.BindingCheck;
+import com.dwarfeng.subgrade.sdk.interceptor.login.LoginRequired;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.multipart.MultipartHttpServletRequest;
+import org.springframework.web.multipart.commons.CommonsMultipartResolver;
+
+import javax.servlet.http.HttpServletRequest;
+import java.io.ByteArrayOutputStream;
+import java.io.InputStream;
+import java.nio.charset.StandardCharsets;
+import java.util.Objects;
+
+/**
+ * 头像控制器。
+ *
+ * @author DwArFeng
+ * @since 1.0.1
+ */
+@RestController
+@RequestMapping("/api/v1/clannad")
+public class AvatarController {
+
+    private final AvatarResponseService service;
+
+    private final CommonsMultipartResolver commonsMultipartResolver;
+
+    private final ServiceExceptionMapper sem;
+
+    public AvatarController(
+            AvatarResponseService service,
+            CommonsMultipartResolver commonsMultipartResolver,
+            ServiceExceptionMapper sem
+    ) {
+        this.service = service;
+        this.commonsMultipartResolver = commonsMultipartResolver;
+        this.sem = sem;
+    }
+
+    @GetMapping("/avatar/{stringId}/exists")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<Boolean> exists(HttpServletRequest request, @PathVariable("stringId") String stringId) {
+        try {
+            boolean exists = service.exists(new StringIdKey(stringId));
+            return FastJsonResponseData.of(ResponseDataUtil.good(exists));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(Boolean.class, e, sem));
+        }
+    }
+
+    @GetMapping("/avatar/{stringId}")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<FastJsonAvatarInfo> get(
+            HttpServletRequest request, @PathVariable("stringId") String stringId
+    ) {
+        try {
+            AvatarInfo avatarInfo = service.get(new StringIdKey(stringId));
+            return FastJsonResponseData.of(ResponseDataUtil.good(FastJsonAvatarInfo.of(avatarInfo)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(FastJsonAvatarInfo.class, e, sem));
+        }
+    }
+
+    @GetMapping("/avatar/{stringId}/download")
+    @BehaviorAnalyse
+    @BindingCheck
+    @LoginRequired
+    public ResponseEntity<Object> downloadAvatar(
+            HttpServletRequest request, @PathVariable("stringId") String stringId
+    ) {
+        HttpHeaders headers = new HttpHeaders();
+        Object body;
+        try {
+            AvatarDownloadInfo avatarDownloadInfo = service.downloadAvatar(new StringIdKey(stringId));
+            // 将文件名转换成 HTTP 标准文件名编码下的格式。
+            String fileName = adjustFileNameEncoding(avatarDownloadInfo.getOriginName());
+            headers.add("Content-Disposition", "attachment;filename=" + fileName);
+            body = avatarDownloadInfo.getContent();
+        } catch (Exception e) {
+            body = FastJsonResponseData.of(ResponseDataUtil.bad(Object.class, e, sem));
+        }
+        return new ResponseEntity<>(body, headers, HttpStatus.OK);
+    }
+
+    @PostMapping("/avatar/{stringId}/upload")
+    @BehaviorAnalyse
+    @BindingCheck
+    @LoginRequired
+    public FastJsonResponseData<Object> uploadAvatar(
+            HttpServletRequest request, @PathVariable("stringId") String stringId
+    ) {
+        try {
+            // 确认请求合法。
+            if (!commonsMultipartResolver.isMultipart(request)) {
+                throw new IllegalStateException("请求不是标准的文件上传请求");
+            }
+
+            //获取 multiRequest 中的文件。
+            MultipartHttpServletRequest multipartHttpServletRequest = commonsMultipartResolver.resolveMultipart(request);
+            MultipartFile file = multipartHttpServletRequest.getFile("file");
+            if (Objects.isNull(file)) {
+                throw new IllegalStateException("请求体中缺少 file 属性");
+            }
+
+            // 解析文件内容。
+            String originFileName = file.getOriginalFilename();
+            byte[] content;
+            try (InputStream in = file.getInputStream(); ByteArrayOutputStream bout = new ByteArrayOutputStream()) {
+                IOUtil.trans(in, bout, 4096);
+                bout.flush();
+                content = bout.toByteArray();
+            }
+
+            // 将文件内容转换为接口需要的格式，并上传。
+            service.uploadAvatar(new StringIdKey(stringId), new AvatarUploadInfo(originFileName, content));
+
+            // 返回响应结果。
+            return FastJsonResponseData.of(ResponseDataUtil.good(null));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(Object.class, e, sem));
+        }
+    }
+
+    private String adjustFileNameEncoding(String fileName) {
+        return new String(fileName.getBytes(), StandardCharsets.ISO_8859_1);
+    }
+}